+++ example/collection_api.d
@@ -6,18 +6,23 @@ import std.stdio;
 void cleanupCollections()
 {
     auto database = new Database();
-    foreach (collection; database.collections)
-        database.deleteCollection(collection.name);
+    foreach (collection; database.collections) {
+        if (collection.name.front != '_')
+            database.deleteCollection(collection.name);
+    }
 }
 
 void main()
 {
     cleanupCollections();
 
+    immutable defaultCollectionNum = 4;
+
     Configuration config;
     auto database = new Database(config);
     {
-        assert(database.collections.length == 0);
+        // _statistics, _trx, _users, _replication
+        assert(database.collections.length == defaultCollectionNum);
     }
 
     immutable name = "test_collection";
@@ -33,7 +38,7 @@ void main()
         assert(created.isLoaded);
         assert(!created.property.isVolatile);
         assert(created.property.waitForSync);
-        assert(database.collections.length == 1);
+        assert(database.collections.length == 1 + defaultCollectionNum);
     }
 
     immutable nameVolatile = "test_volatile";
@@ -49,7 +54,7 @@ void main()
         assert(volatiled.isLoaded);
         assert(volatiled.property.isVolatile);
         assert(!volatiled.property.waitForSync);
-        assert(database.collections.length == 2);
+        assert(database.collections.length == 2 + defaultCollectionNum);
     }
 
     writeln("Get '", name, "' collection");
